JAMS Documentation

JAMS Documentation

Index

JAMS.Euler_integrator — Method

Euler_integrator(system, dt, t_stop; kwargs...) -> SIM

Evolves the initial state defined in system according to the forces and DOFevolvers in it, using the Euler(-Maruyama) algorithm. Returns SIM struct with final states for chaining simulations.

Arguments

system: Instance of System struct

dt: Integration time step size

t_stop: Integration end time

Keywords

seed: master seed for reproducibility (Default: nothing)

Tsave: Save every Tsave dt-steps (Default: nothing)

save_functions: Tuple of functions to save data (Default: nothing)

save_folder_path: Folder path to save data (need not exist) (Default: nothing)

save_tag: Add custom name to save file name, so that it becomes "savetag.rawdata.h5" (Default: nothing)

Tplot: Plot every Tplot dt-steps (Default: nothing)

fps: Default frames per second for rendering of live plotting (Default: 30)

plot_functions: Tuple of plot function to live plot (Default: nothing)

plotdim: Default plotting dimension, change to 3 for 3d plot (Default: 2)

record_folder_path: Folder path to save live recording in, need not exist (Default: nothing)

crf: Default compression rate for saving the live recording (Default: 23)

res: Pixel resoluation of plotting, e.g. (1000,1000) (Default: nothing)

format: Default format for saving live recording (Default: "mp4")

sbs: Set to true for side-by-side plotting, to be used i.c.m. with plotdim=3 (Default: false)

Index

JAMS.Euler_integrator